refactor: reorganize Tool serde utility functions
