+++ haystack/tools/__init__.py
@@ -6,17 +6,18 @@
 
 # ruff: noqa: I001 (ignore import order as we need to import Tool before ComponentTool)
 from .from_function import create_tool_from_function, tool
-from .tool import Tool, _check_duplicate_tool_names, deserialize_tools_inplace
+from .tool import Tool, _check_duplicate_tool_names
 from .component_tool import ComponentTool
+from .serde_utils import deserialize_tools_inplace, serialize_tools_or_toolset
 from .toolset import Toolset
 
-
 __all__ = [
-    "Tool",
     "_check_duplicate_tool_names",
-    "deserialize_tools_inplace",
-    "create_tool_from_function",
-    "tool",
     "ComponentTool",
+    "create_tool_from_function",
+    "deserialize_tools_inplace",
+    "serialize_tools_or_toolset",
+    "Tool",
     "Toolset",
+    "tool",
 ]
